:sparkles: Implementação das DTOs na Service e Controller de Fornecedor, para melhor tratamento FrontEnd

+++ src/main/java/com/industria/cafeeira/model/service/FornecedorService.java
@@ -1,8 +1,11 @@
 package com.industria.cafeeira.model.service;
 
+import com.industria.cafeeira.model.dto.FornecedorRequestDTO;
 import com.industria.cafeeira.model.entities.Fornecedor;
 import com.industria.cafeeira.model.entities.Lote;
 import com.industria.cafeeira.model.repository.FornecedorRepository;
+import com.industria.cafeeira.view.FornecedorMapper;
+import com.industria.cafeeira.view.FornecedorResponseDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,42 +17,50 @@ import java.util.List;
 public class FornecedorService {
     @Autowired
     private FornecedorRepository repository;
+    @Autowired
+    private FornecedorMapper mapper;
 
-    public List<Fornecedor>findAll(){
-        return repository.findAll();
+    public List<FornecedorResponseDTO>findAll(){
+       return repository.findAll().stream().map(mapper::toResponse).toList();
     }
 
-    public Fornecedor findById(Long id){
-        return repository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Fornecedor não encontrado!"));
+    public FornecedorResponseDTO findById(Long id){
+       Fornecedor fornecedor = repository.findById(id)
+               .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Fornecedor não encontrado!"));
+       return mapper.toResponse(fornecedor);
     }
 
-    public Fornecedor create(Fornecedor fornecedor){
+    public FornecedorResponseDTO create(FornecedorRequestDTO req){
+       Fornecedor fornecedor = mapper.toEntity(req);
+
+        fornecedor.setCodigo("COD-" + System.currentTimeMillis());
+
         Fornecedor salvo = repository.save(fornecedor);
 
-        String codigoGerado = "COD-" + String.format("%06d", salvo.getId());
-
-        salvo.setCodigo(codigoGerado);
-
-        return repository.save(salvo);
+        return mapper.toResponse(salvo);
     }
 
-    public Fornecedor update(Long id, Fornecedor dados){
-        Fornecedor fornecedor = findById(id);
+    public FornecedorResponseDTO update(Long id, FornecedorRequestDTO dados){
+      Fornecedor existente = repository.findById(id)
+              .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Fornecedor não encontrado!"));
 
-        //CRUD simples, service recebe a entity diretamente
-        fornecedor.setNome_fantasia(dados.getNome_fantasia());
-        fornecedor.setRazao_social(dados.getRazao_social());
-        fornecedor.setCpf_cnpj(dados.getCpf_cnpj());
-        fornecedor.setInscricao_estadual(dados.getInscricao_estadual());
-        fornecedor.setInscricao_municipal(dados.getInscricao_municipal());
-        fornecedor.setLogradouro(dados.getLogradouro());
-        fornecedor.setSimples_nacional(dados.isSimples_nacional());
+      existente.setNome_fantasia(dados.nome_fantasia());
+      existente.setRazao_social(dados.razao_social());
+      existente.setCpf_cnpj(dados.cpf_cnpj());
+      existente.setInscricao_estadual(dados.inscricao_estadual());
+      existente.setInscricao_municipal(dados.inscricao_municipal());
+      existente.setLogradouro(dados.logradouro());
+      existente.setSimples_nacional(dados.simples_nacional());
 
-        return repository.save(fornecedor);
+      Fornecedor salvo = repository.save(existente);
+
+      return mapper.toResponse(salvo);
     }
 
     public void delete(Long id){
+        if (!repository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Fornecedor não encontrado!");
+        }
         repository.deleteById(id);
     }
 }
